+++ src/state/mod.rs
@@ -3,5 +3,7 @@ pub mod chainstate;
 pub mod context;
 
 pub use chain::{Chain, ChainIterator};
-pub use chainstate::{ChainstateManager, ChainstateManagerBuilder, ProcessBlockResult};
+pub use chainstate::{
+    ChainstateManager, ChainstateManagerBuilder, ProcessBlockHeaderResult, ProcessBlockResult,
+};
 pub use context::{ChainParams, ChainType, Context, ContextBuilder};